again modified login and signup page
